+++ src/LIAnsureProtect.Application/DependencyInjection.cs
@@ -1,12 +1,21 @@
+using FluentValidation;
+using LIAnsureProtect.Application.Common.Behaviors;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace LIAnsureProtect.Application;
-
 
 public static class DependencyInjection
 {
     public static IServiceCollection AddApplication(this IServiceCollection services)
     {
+        services.AddMediatR(configuration =>
+        {
+            configuration.RegisterServicesFromAssembly(typeof(DependencyInjection).Assembly);
+            configuration.AddOpenBehavior(typeof(ValidationBehavior<,>));
+        });
+
+        services.AddValidatorsFromAssembly(typeof(DependencyInjection).Assembly);
+
         return services;
     }
 }